brungasinc_project/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cloudinary==1.26.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Deprecated==1.2.12
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cloudinary-storage==0.3.0
django-environ==0.4.5
django-heroku==0.3.1
django-oauth-toolkit==1.5.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
heroku3==4.2.3
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tox==3.24.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
whitenoise==5.3.0
wrapt==1.12.1